+++ app/models/maestrano/connector/rails/concerns/connector_logger.rb
@@ -2,8 +2,10 @@ module Maestrano::Connector::Rails::Concerns::ConnectorLogger
   extend ActiveSupport::Concern
 
   module ClassMethods
-    def log(level, organization, msg)
-      Rails.logger.method(level).call("uid=#{organization&.uid}, org_uid= #{organization&.org_uid} , tenant=#{organization&.tenant}): #{msg}")
+    def log(level, organization, msg, params = {})
+      line = "uid=\"#{organization&.uid}\", org_uid=\"#{organization&.org_uid}\", tenant=\"#{organization&.tenant}\""
+      line = "#{line}, #{params.map { |k, v| "#{k}=\"#{v}\"" }.join(', ')}" unless params.blank?
+      Rails.logger.method(level).call("#{line}, message=\"#{msg}\"")
     end
   end
 end